Refactoring find user, moving request and response files to another folder
