+++ src/model/user_domain.go
@@ -1,21 +1,11 @@
 package model
 
-import "encoding/json"
-
 type userDomain struct {
 	id        string
 	firstName string
 	lastName  string
 	email     string
 	age       int8
-}
-
-func (ud *userDomain) ToJSON() (string, error) {
-	output, err := json.Marshal(ud)
-	if err != nil {
-		return "", err
-	}
-	return string(output), nil
 }
 
 func (ud *userDomain) GetEmail() string {
