+++ src/models/user_model.go
@@ -1,0 +1,11 @@
+package models
+
+import "go.mongodb.org/mongo-driver/bson/primitive"
+
+type UserModel struct {
+	Id        primitive.ObjectID `json:"id,omitempty" bson:"_id"`
+	Email     string             `json:"email,omitempty"`
+	FirstName string             `json:"firstName,omitempty"`
+	LastName  string             `json:"lastName,omitempty"`
+	Age       int8               `json:"age,omitempty"`
+}
