+++ src/models/user_response.go
@@ -1,7 +1,7 @@
-package response
+package models
 
 type UserResponse struct {
-	Id        string `json:"id,omitempty"`
+	Id        string `json:"id,omitempty" bson:"_id"`
 	FirstName string `json:"firstName,omitempty"`
 	LastName  string `json:"lastName,omitempty"`
 	Email     string `json:"email,omitempty"`
