+++ src/repository/find_user_repository.go
@@ -5,24 +5,22 @@ import (
 	"errors"
 	"fmt"
 	"github.com/lucasolsi-wex/go-crud/src/config/custom_errors"
-	"github.com/lucasolsi-wex/go-crud/src/model"
-	"github.com/lucasolsi-wex/go-crud/src/model/converter"
-	"github.com/lucasolsi-wex/go-crud/src/model/entity"
+	"github.com/lucasolsi-wex/go-crud/src/models"
 	"github.com/spf13/viper"
 	"go.mongodb.org/mongo-driver/bson"
 	"go.mongodb.org/mongo-driver/bson/primitive"
 	"go.mongodb.org/mongo-driver/mongo"
 )
 
-func (userRepo *userRepository) FindUserById(id string) (model.UserDomainInterface, *custom_errors.CustomErr) {
+func (userRepo *userRepository) FindUserById(id string) (*models.UserResponse, *custom_errors.CustomErr) {
 	collectionName := viper.GetString(MongoDBUserDb)
 	collection := userRepo.databaseConnection.Collection(collectionName)
 
-	userEntity := &entity.UserEntity{}
+	userResponse := &models.UserResponse{}
 
 	objectId, _ := primitive.ObjectIDFromHex(id)
 	filter := bson.D{{Key: "_id", Value: objectId}}
-	err := collection.FindOne(context.Background(), filter).Decode(userEntity)
+	err := collection.FindOne(context.Background(), filter).Decode(userResponse)
 
 	if err != nil {
 		if errors.Is(err, mongo.ErrNoDocuments) {
@@ -34,5 +32,5 @@ func (userRepo *userRepository) FindUserById(id string) (model.UserDomainInterfa
 		return nil, custom_errors.NewInternalServerError(errorMessage)
 	}
 
-	return converter.ConvertEntityToDomain(*userEntity), nil
+	return userResponse, nil
 }
