+++ src/view/convert_domain_to_response.go
@@ -2,11 +2,11 @@ package view
 
 import (
 	"github.com/lucasolsi-wex/go-crud/src/model"
-	"github.com/lucasolsi-wex/go-crud/src/model/response"
+	"github.com/lucasolsi-wex/go-crud/src/models"
 )
 
-func ConvertDomainToResponse(userDomain model.UserDomainInterface) response.UserResponse {
-	return response.UserResponse{
+func ConvertDomainToResponse(userDomain model.UserDomainInterface) models.UserResponse {
+	return models.UserResponse{
 		Id:        userDomain.GetId(),
 		FirstName: userDomain.GetFirstName(),
 		LastName:  userDomain.GetLastName(),
